docs(autodoc) remove unused autodoc/conf scripts
Related with #8062

This change should (probably) not be merged in the EE branch.

The docs generation for the conf file of the opensource version
is now being produced from a script in the Enterprise repo which
will produce both versions of the scripts (OSS & Enterprise).

As a result this code is dead code on the OSS repo.

Merging this and #8092 upstream might have unintended consequences if
the scripts used there for autodocs are "descendants" of these scripts.